pyproject.toml
[tool.poetry]
name = "Passsssword"
version = "0.1.0"
description = "Utilize 1Password vault data with Python decorators and context managers. Driven by the 1Password CLI interface. 🐍🔐"
authors = ["Chris Lemke <1@lemke.ai>"]
license = "MIT"
readme = "docs/README.md"


[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^1.0.0"

[tool.poetry.group.develop]
optional = true

[tool.poetry.group.develop.dependencies]
black = "^23.10.1"
mypy = "^1.6.1"
ruff = "^0.1.3"

[tool.poetry.group.tests]
optional = true

[tool.poetry.group.tests.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
disallow_subclassing_any = false
ignore_missing_imports = true
disallow_any_generics = false
warn_return_any = false
warn_unused_ignores = false
no_namespace_packages = true


[tool.ruff]
ignore = ["E501", "E402"]
line-length = 88
target-version = "py311"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
